Fix: nickname, age는 카카오에서 받아 온 정보 그대로 넘기기 & 그 외 정보는 patch로 가능하게끔 수정

+++ letsit_backend/letsit_backend/src/main/java/letsit_backend/dto/profile/ProfileRequestDto.java
@@ -15,7 +15,7 @@ public class ProfileRequestDto {
     private Long userId;
     private Long profileId;
     private Profile.Manner_tier mannerTier;
-    private int mannerScore;
+    private double mannerScore;
     private String nickname;
     private String age;
     private String sns;
